refactor(emitter): split lowering import usage
