+++ crates/tsz-emitter/src/lowering/mod.rs
@@ -42,6 +42,7 @@ mod decorator_helpers;
 mod helpers;
 mod helpers_class_expr_static_name;
 mod helpers_private_fields;
+mod import_usage;
 mod name_helpers;
 mod parameter_prologue;
 mod visit_children;
